refactor (LottoNumbers) : 기본 로또 번호들 LottoNumbers 클래스로 분리
