+++ src/main/java/lotto/LottoNumber.java
@@ -1,24 +1,13 @@
 package lotto;
 
-import java.util.List;
 import java.util.Objects;
-import java.util.stream.Collectors;
-import java.util.stream.IntStream;
 
 public class LottoNumber implements Comparable {
 
     public static final int LOTTO_NUMBER_MIN = 1;
     public static final int LOTTO_NUMBER_MAX = 45;
 
-    private static final List<Integer> defaultLottoNumbers;
-
     private int number;
-
-    static {
-        defaultLottoNumbers = IntStream.range(LOTTO_NUMBER_MIN, LOTTO_NUMBER_MAX + 1)
-                .boxed()
-                .collect(Collectors.toList());
-    }
 
     public LottoNumber(int number) {
         if (!isValidNumber(number)) {
@@ -35,19 +24,13 @@ public class LottoNumber implements Comparable {
         return number;
     }
 
-    public static List<LottoNumber> getDefaultLottoNumbers() {
-        return defaultLottoNumbers.stream()
-                .map(LottoNumber::of)
-                .collect(Collectors.toList());
-    }
-
     public static boolean isValidNumber(int number) {
         return number >= LOTTO_NUMBER_MIN && number <= LOTTO_NUMBER_MAX;
     }
 
     @Override
     public int compareTo(Object o) {
-        return Integer.compare(number, ((LottoNumber)o).getNumber());
+        return Integer.compare(number, ((LottoNumber) o).getNumber());
     }
 
     @Override
