+++ src/test/java/lotto/LottoNumberTest.java
@@ -23,12 +23,6 @@ public class LottoNumberTest {
     }
 
     @Test
-    @DisplayName("로또 숫자 리스트의 크기를 검증한다.")
-    void 로또_숫자_리스트_테스트() {
-        assertThat(LottoNumber.getDefaultLottoNumbers()).hasSize(LottoNumber.LOTTO_NUMBER_MAX);
-    }
-
-    @Test
     @DisplayName("로또 번호 검증 메소드 테스트")
     void 로또_번호_검증_테스트() {
         assertThat(LottoNumber.isValidNumber(1)).isTrue();
